+++ src/OpenGL/ShaderOGLLine.cpp
@@ -42,11 +42,11 @@ void ShaderOGLLine::render(Point2D a, Point2D b, Color c, float weight, GLuint f
 
     LoadQuadMatrixUniforms();
 
-    const auto normAx = (float) ((float)a.y / (float) Config::get()->screenHeight);
-    const auto normAy = (float) ((float)a.x / (float) Config::get()->screenWidth);
+    const auto normAx = (float) a.y / (float) Config::get()->screenHeight;
+    const auto normAy = (float) a.x / (float) Config::get()->screenWidth;
 
-    const auto normBx = (float) ((float)b.y / (float) Config::get()->screenHeight);
-    const auto normBy = (float) ((float)b.x / (float) Config::get()->screenWidth);
+    const auto normBx = (float) b.y / (float) Config::get()->screenHeight;
+    const auto normBy = (float) b.x / (float) Config::get()->screenWidth;
 
     setVec2Uniform(lineStartUniform, glm::vec2(normAx, normAy));
     setVec2Uniform(lineEndUniform, glm::vec2(normBx, normBy));
